requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic
pydantic[email]
flet
httpx
passlib[bcrypt]
PyJWT
